Moving commands to the testing area, soon they will be moved to their own packages.
