+++ main.js
@@ -1,5 +1,4 @@
 const Command = require('./src/command');
-const Commands = require('./src/commands');
 const CommandEvent = require('./src/events/command-event');
 const CommandInput = require('./src/command-input');
 const CommandInputEvent = require('./src/events/command-input-event');
@@ -7,7 +6,6 @@ const CommandInputValidationEvent = require('./src/events/command-input-validati
 const Dispatcher = require('./src/dispatcher');
 const Process = require('./src/process');
 const Events = require('./src/events');
-
 
 if (require.resolve('data-sanitizers')) {
     require('./src/subscribers/sanitation');
@@ -23,7 +21,6 @@ module.exports = {
     CommandInput,
     CommandInputEvent,
     CommandInputValidationEvent,
-    Commands,
     Dispatcher,
     Events,
     Process,
