--- src/commands.js
@@ -1,9 +1,0 @@
-const Math = require('./commands/math');
-const Primitives = require('./commands/primitives');
-const Dispatcher = require('./commands/dispatcher');
-
-module.exports = {
-    Math,
-    Primitives,
-    Dispatcher
-};
