+++ test/commands/dispatcher.js
