+++ test/commands/dispatcher/reset.js
@@ -1,4 +1,4 @@
-const Command = require('../../command');
+const Command = require('../../../src/command');
 
 class ResetCommand extends Command {
     /**
