+++ test/commands/dispatcher/stop-processing.js
@@ -1,4 +1,4 @@
-const Command = require('../../command');
+const Command = require('../../../src/command');
 
 class StopProcessingCommand extends Command {
     /**
