+++ test/commands/math.js
