+++ test/commands/primitives.js
