+++ test/commands/primitives/string.js
@@ -1,51 +1,38 @@
-const { assert, expect } = require('chai');
+const Command = require('../../../src/command');
 const Dispatcher = require('../../../src/dispatcher');
-const StringCommand = require('../../../src/commands/primitives/string');
 
-describe('String Command', () => {
-    describe('process(dispatcher)', () => {
-        it ('passes when result is object.', async () => {
-            let string = new StringCommand([], {
-                value: 'Hello World'
-            });
+class StringCommand extends Command {
+    static describe() {
+        return {
+            key: 'string',
+            name: 'String',
+            description: 'Create a text value.',
+            options: [
+                {
+                    name: 'value',
+                    type: 'string',
+                    description: 'The text that you want to create.'
+                }
+            ]
+        };
+    }
 
-            let dispatcher = new Dispatcher(this).startProcessing();
-            let results    = await string.process(dispatcher);
+    static defaults() {
+        return { value: '' };
+    }
 
-            assert.isObject(results);
-            assert.hasAllKeys(results, ['value']);
-            assert.isString(results.value);
-            assert.equal(results.value, 'Hello World');
-        });
-    });
+    /**
+     * Returns the string value.
+     *
+     * @param {Dispatcher} dispatcher
+     * @returns {{value: string}}
+     */
+    execute(dispatcher) {
+        let value = this.options.get('value');
+        return {
+            value: value
+        }
+    }
+}
 
-    describe('execute(dispatcher)', () => {
-        it ('passes when result is object.', async () => {
-            let string = new StringCommand([], {
-                value: 'Hello World'
-            });
-
-            let dispatcher = new Dispatcher(this).startProcessing();
-            let results    = await string.execute(dispatcher);
-
-            assert.isObject(results);
-            assert.hasAllKeys(results, ['value']);
-            assert.isString(results.value);
-            assert.equal(results.value, 'Hello World');
-        });
-    });
-/*
-    describe('execute(dispatcher)', () => {
-        it ('fails when value is not a string.', async () => {
-            let string = new StringCommand([], {
-                value: 123
-            });
-
-            let dispatcher = new Dispatcher(this).startProcessing();
-            let results = await string.execute(dispatcher);
-
-            assert.isString(results.value);
-        });
-    });
-*/
-});
+module.exports = StringCommand;
